move: UserInfoKey enum을 따로 파일 분리
