+++ DevFriends/DevFriends/Data/UserInfoKey.swift
@@ -1,0 +1,19 @@
+//
+//  UserInfoKey.swift
+//  DevFriends
+//
+//  Created by 유승원 on 2022/12/04.
+//
+
+import Foundation
+
+enum UserInfoKey: String {
+    case uid = "UID"
+    case email = "EMAIL"
+    case nickname = "NICKNAME"
+    case job = "JOB"
+    case profile = "PROFILE"
+    case categoryIDs = "CATEGORYIDS"
+    case appliedGroupIDs = "APPLIEDGROUPIDS"
+    case isEnabledAutoLogin = "ISENABLEDAUTOLOGIN"
+}
